refactor: move docker image cleanup from apps plugin to builder plugin
This isn't where it belongs - images should be cleaned up in a builder plugin (or the builders plugin) and not in the apps plugin, which has nothing to do with docker images anyways.

Also remove the code from the scheduler plugin, as that will only do anything if the scheduler is enabled for the app.
